requirements.txt
markdownify==0.10.0
pandas==1.1.0
PyYAML==6.0.0
